+++ libraries/kotlinx-metadata/jvm/test/kotlinx/metadata/test/MetadataSmokeTest.kt
@@ -41,9 +41,9 @@ class MetadataSmokeTest {
                         if (type != JvmFunctionExtensionVisitor.TYPE) return null
 
                         return object : JvmFunctionExtensionVisitor() {
-                            override fun visit(desc: JvmMethodSignature?) {
-                                if (Flag.Function.IS_INLINE(flags) && desc != null) {
-                                    inlineFunctions += desc.asString()
+                            override fun visit(signature: JvmMethodSignature?) {
+                                if (Flag.Function.IS_INLINE(flags) && signature != null) {
+                                    inlineFunctions += signature.asString()
                                 }
                             }
                         }